+++ qrinput.c
@@ -430,21 +430,21 @@ static int QRinput_encodeModeNum(QRinput_List *entry, BitStream *bstream, int ve
 
 	words = entry->size / 3;
 	for(i = 0; i < words; i++) {
-		val  = (entry->data[i*3  ] - '0') * 100;
-		val += (entry->data[i*3+1] - '0') * 10;
-		val += (entry->data[i*3+2] - '0');
+		val  = (unsigned int)(entry->data[i*3  ] - '0') * 100;
+		val += (unsigned int)(entry->data[i*3+1] - '0') * 10;
+		val += (unsigned int)(entry->data[i*3+2] - '0');
 
 		ret = BitStream_appendNum(bstream, 10, val);
 		if(ret < 0) return -1;
 	}
 
 	if(entry->size - words * 3 == 1) {
-		val = entry->data[words*3] - '0';
+		val = (unsigned int)(entry->data[words*3] - '0');
 		ret = BitStream_appendNum(bstream, 4, val);
 		if(ret < 0) return -1;
 	} else if(entry->size - words * 3 == 2) {
-		val  = (entry->data[words*3  ] - '0') * 10;
-		val += (entry->data[words*3+1] - '0');
+		val  = (unsigned int)(entry->data[words*3  ] - '0') * 10;
+		val += (unsigned int)(entry->data[words*3+1] - '0');
 		BitStream_appendNum(bstream, 7, val);
 		if(ret < 0) return -1;
 	}
@@ -715,9 +715,9 @@ static int QRinput_encodeModeStructure(QRinput_List *entry, BitStream *bstream, 
 
 	ret = BitStream_appendNum(bstream, 4, QRSPEC_MODEID_STRUCTURE);
 	if(ret < 0) return -1;
-	ret = BitStream_appendNum(bstream, 4, entry->data[1] - 1);
+	ret = BitStream_appendNum(bstream, 4, entry->data[1] - 1U);
 	if(ret < 0) return -1;
-	ret = BitStream_appendNum(bstream, 4, entry->data[0] - 1);
+	ret = BitStream_appendNum(bstream, 4, entry->data[0] - 1U);
 	if(ret < 0) return -1;
 	ret = BitStream_appendNum(bstream, 8, entry->data[2]);
 	if(ret < 0) return -1;
